requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
HTMLParser==0.0.2
idna==3.3
python-dotenv==0.19.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
